Fixed argument types to match format strings.

+++ libcpluff/ploader.c
@@ -165,15 +165,15 @@ static void descriptor_errorf(ploader_context_t *plcontext, int warn,
 		cpi_warnf(plcontext->context,
 			N_("Suspicious plug-in descriptor content in %s, line %d, column %d (%s)."),
 		plcontext->file,
-		XML_GetCurrentLineNumber(plcontext->parser),
-		XML_GetCurrentColumnNumber(plcontext->parser) + 1,
+		(int) XML_GetCurrentLineNumber(plcontext->parser),
+		(int) XML_GetCurrentColumnNumber(plcontext->parser) + 1,
 		message);
 	} else {				
 		cpi_errorf(plcontext->context,
 			N_("Invalid plug-in descriptor content in %s, line %d, column %d (%s)."),
 			plcontext->file,
-			XML_GetCurrentLineNumber(plcontext->parser),
-			XML_GetCurrentColumnNumber(plcontext->parser) + 1,
+			(int) XML_GetCurrentLineNumber(plcontext->parser),
+			(int) XML_GetCurrentColumnNumber(plcontext->parser) + 1,
 			message);
 	}
 	if (!warn) {
@@ -192,8 +192,8 @@ static void resource_error(ploader_context_t *plcontext) {
 		cpi_errorf(plcontext->context,
 			N_("Insufficient system resources to parse plug-in descriptor content in %s, line %d, column %d."),
 			plcontext->file,
-			XML_GetCurrentLineNumber(plcontext->parser),
-			XML_GetCurrentColumnNumber(plcontext->parser) + 1);
+			(int) XML_GetCurrentLineNumber(plcontext->parser),
+			(int) XML_GetCurrentColumnNumber(plcontext->parser) + 1);
 	}
 	plcontext->resource_error_count++;
 }
@@ -1088,8 +1088,8 @@ CP_C_API cp_plugin_info_t * cp_load_plugin_descriptor(cp_context_t *context, con
 				cpi_errorf(context,
 					N_("XML parsing error in %s, line %d, column %d (%s)."),
 					file,
-					XML_GetErrorLineNumber(parser),
-					XML_GetErrorColumnNumber(parser) + 1,
+					(int) XML_GetErrorLineNumber(parser),
+					(int) XML_GetErrorColumnNumber(parser) + 1,
 					XML_ErrorString(XML_GetErrorCode(parser)));
 				cpi_unlock_context(context);
 			}
@@ -1274,8 +1274,8 @@ CP_C_API cp_plugin_info_t * cp_load_plugin_descriptor_from_memory(cp_context_t *
 			  cpi_errorf(context,
 				  N_("XML parsing error in %s, line %d, column %d (%s)."),
 				  file,
-				  XML_GetErrorLineNumber(parser),
-				  XML_GetErrorColumnNumber(parser) + 1,
+				  (int) XML_GetErrorLineNumber(parser),
+				  (int) XML_GetErrorColumnNumber(parser) + 1,
 				  XML_ErrorString(XML_GetErrorCode(parser)));
 			  cpi_unlock_context(context);
 		  }
